service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
google-cloud-firestore==2.13.1
python-dotenv==1.0.0
firebase-admin===6.3.0
pytest==7.4.0
